Big changes.
Decided to implement a priority queue data structure. It seemed reasonable to group it together with the graph data structure in a new module. Also starting to group the search algorithms in a more structured way, to keep the testing infrastructure passing I had to change some of the test modules.